backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
celery>=5.3
redis>=4.6
django-celery-results>=2.5
python-dotenv>=1.0
